src/requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
flake8==5.0.4
mccabe==0.7.0
Pillow==9.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.5
sqlparse==0.4.3
